+++ intelmq/lib/upgrades.py
@@ -21,7 +21,7 @@ __all__ = ['v100_dev7_modify_syntax',
            ]
 
 
-def v201_defaults_statistics():
+def v200_defaults_statistics(defaults, runtime, dry_run):
     """
     Inserting "statistics_*" parameters into defaults.conf file
     """
@@ -30,31 +30,19 @@ def v201_defaults_statistics():
               "statistics_password": None,
               "statistics_port": 6379
               }
-    changed = False
-    defaults = utils.load_configuration(intelmq.DEFAULTS_CONF_FILE)
+    changed = None
     for key, value in values.items():
         if key not in defaults:
             defaults[key] = value
             changed = True
-    if not changed:
-        # all keys exist already
-        return None
-
-    try:
-        with open(intelmq.DEFAULTS_CONF_FILE, 'w') as handle:
-            json.dump(defaults, fp=handle, indent=4, sort_keys=True,
-                      separators=(',', ': '))
-    except PermissionError:
-        return 'Can\'t update defaults configuration: Permission denied.'
-    return True
+    return changed, defaults, runtime
 
 
-def v201_defaults_broker():
+def v200_defaults_broker(defaults, runtime, dry_run):
     """
     Inserting "*_pipeline_broker" and deleting broker into/from defaults configuration
     """
-    changed = False
-    defaults = utils.load_configuration(intelmq.DEFAULTS_CONF_FILE)
+    changed = None
     values = {"destination_pipeline_broker": defaults.get("broker", "redis"),
               "source_pipeline_broker": defaults.get("broker", "redis"),
               }
@@ -65,42 +53,25 @@ def v201_defaults_broker():
     if "broker" in defaults:
         del defaults["broker"]
         changed = True
-    if not changed:
-        # all keys exist already and broker is gone
-        return None
 
-    try:
-        with open(intelmq.DEFAULTS_CONF_FILE, 'w') as handle:
-            json.dump(defaults, fp=handle, indent=4, sort_keys=True,
-                      separators=(',', ': '))
-    except PermissionError:
-        return 'Can\'t update defaults configuration: Permission denied.'
-    return True
+    return changed, defaults, runtime
 
 
-def v112_feodo_tracker_ips():
+def v112_feodo_tracker_ips(defaults, runtime, dry_run):
     """
     Fix URL of feodotracker IPs feed in runtime configuration
     """
-    runtime = utils.load_configuration(intelmq.RUNTIME_CONF_FILE)
-    changed = False
+    changed = None
     for bot_id, bot in runtime.items():
         if bot["parameters"].get("http_url") == "https://feodotracker.abuse.ch/blocklist/?download=ipblocklist":
             bot["parameters"]["http_url"] = "https://feodotracker.abuse.ch/downloads/ipblocklist.csv"
             changed = True
-    if not changed:
-        return None
+    changed = True
 
-    try:
-        with open(intelmq.RUNTIME_CONF_FILE, 'w') as handle:
-            json.dump(runtime, fp=handle, indent=4, sort_keys=True,
-                      separators=(',', ': '))
-    except PermissionError:
-        return 'Can\'t update runtime configuration: Permission denied.'
-    return True
+    return changed, defaults, runtime
 
 
-def v112_feodo_tracker_domains():
+def v112_feodo_tracker_domains(defaults, runtime, dry_run):
     """
     Search for discontinued feodotracker domains feed
     """
@@ -111,13 +82,14 @@ def v112_feodo_tracker_domains():
             found = bot_id
 
     if not found:
-        return None
+        return None, defaults, runtime
     else:
         return ('The discontinued feed "Feodo Tracker Domains" has been found '
-                'as bot %r. Remove it yourself please.' % found)
+                'as bot %r. Remove it yourself please.' % found,
+                defaults, runtime)
 
 
-def v110_shadowserver_feednames():
+def v110_shadowserver_feednames(defaults, runtime, dry_run):
     """
     Replace deprecated Shadowserver feednames
     """
@@ -128,26 +100,17 @@ def v110_shadowserver_feednames():
         "Ssl-Freak-Scan": "SSL-FREAK-Vulnerable-Servers",
         "Ssl-Scan": "SSL-POODLE-Vulnerable-Servers",
     }
-    runtime = utils.load_configuration(intelmq.RUNTIME_CONF_FILE)
-    changed = False
+    changed = None
     for bot_id, bot in runtime.items():
         if bot["module"] == "intelmq.bots.parsers.shadowserver.parser":
             if bot["parameters"]["feedname"] in mapping:
                 changed = True
                 bot["parameters"]["feedname"] = mapping[bot["parameters"]["feedname"]]
-    if not changed:
-        return None
 
-    try:
-        with open(intelmq.RUNTIME_CONF_FILE, 'w') as handle:
-            json.dump(runtime, fp=handle, indent=4, sort_keys=True,
-                      separators=(',', ': '))
-    except PermissionError:
-        return 'Can\'t update runtime configuration: Permission denied.'
-    return True
+    return changed, defaults, runtime
 
 
-def v110_deprecations():
+def v110_deprecations(defaults, runtime, dry_run):
     """
     Checking for deprecated runtime configurations
     """
@@ -155,8 +118,7 @@ def v110_deprecations():
         "intelmq.bots.collectors.n6.collector_stomp": "intelmq.bots.collectors.stomp.collector",
         "intelmq.bots.parsers.cymru_full_bogons.parser": "intelmq.bots.parsers.cymru.parser_full_bogons",
     }
-    runtime = utils.load_configuration(intelmq.RUNTIME_CONF_FILE)
-    changed = False
+    changed = None
     for bot_id, bot in runtime.items():
 
         if bot["module"] in mapping:
@@ -204,11 +166,10 @@ def modify_expert_convert_config(old):
     return config
 
 
-def v100_dev7_modify_syntax():
+def v100_dev7_modify_syntax(defaults, runtime, dry_run):
     """
     Migrate modify bot configuration format
     """
-    runtime = utils.load_configuration(intelmq.RUNTIME_CONF_FILE)
     changed = None
     for bot_id, bot in runtime.items():
         if bot["module"] == "intelmq.bots.experts.modify.expert":
@@ -220,12 +181,13 @@ def v100_dev7_modify_syntax():
                         return 'Error converting modify expert syntax. Different size of configurations. Please report this.'
                     changed = True
                     try:
-                        with open(bot["parameters"]["configuration_path"],
-                                  'w') as handle:
-                            json.dump(new_config, fp=handle, indent=4, sort_keys=True,
-                                      separators=(',', ': '))
+                        utils.write_configuration(bot["parameters"]["configuration_path"],
+                                                  new_config)
                     except PermissionError:
-                        return 'Can\'t update %s\'s configuration: Permission denied.' % bot_id
+                        return ('Can\'t update %s\'s configuration: Permission denied.' % bot_id,
+                                defaults, runtime)
+
+    return changed, defaults, runtime
 
     return changed